Move file manager context menus to util
