+++ source/renderer/modules/file-manager/file-item.vue
@@ -112,8 +112,8 @@
   const Sorter = require('./sorter.vue').default
   const formatDate = require('../../../common/util/format-date.js')
   const { trans } = require('../../../common/i18n.js')
-  const fileContextMenu = require('./file-item-context.js')
-  const dirContextMenu = require('./dir-item-context.js')
+  const fileContextMenu = require('./util/file-item-context.js')
+  const dirContextMenu = require('./util/dir-item-context.js')
 
   module.exports = {
     name: 'file-item',
