+++ source/renderer/modules/file-manager/tree-item.vue
@@ -145,8 +145,8 @@
 const findObject = require('../../../common/util/find-object.js')
 const { trans } = require('../../../common/i18n')
 const Sorter = require('./sorter.vue').default
-const fileContextMenu = require('./file-item-context.js')
-const dirContextMenu = require('./dir-item-context.js')
+const fileContextMenu = require('./util/file-item-context.js')
+const dirContextMenu = require('./util/dir-item-context.js')
 
 module.exports = {
   name: 'tree-item',
